restaurant-microservices/services/staff-service/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-filter>=23.5
mysqlclient>=2.2.0
PyJWT>=2.8.0
gunicorn>=21.2.0
